Cleaned up code and renamed socket to images
