+++ backend/routes/images/index.js
